refactor: int 캐스팅을 Math.toIntExact로 교체

+++ src/main/java/aegis/server/domain/member/repository/MemberRecordQueryRepositoryImpl.java
@@ -61,7 +61,7 @@ public class MemberRecordQueryRepositoryImpl implements MemberRecordQueryReposit
             countQuery.setParameter(param.getKey(), param.getValue());
         }
 
-        selectQuery.setFirstResult((int) pageable.getOffset());
+        selectQuery.setFirstResult(Math.toIntExact(pageable.getOffset()));
         selectQuery.setMaxResults(pageable.getPageSize());
 
         List<MemberRecord> content = selectQuery.getResultList();
