+++ src/main/java/aegis/server/domain/payment/repository/TransactionQueryRepositoryImpl.java
@@ -70,7 +70,7 @@ public class TransactionQueryRepositoryImpl implements TransactionQueryRepositor
             countQuery.setParameter(param.getKey(), param.getValue());
         }
 
-        selectQuery.setFirstResult((int) pageable.getOffset());
+        selectQuery.setFirstResult(Math.toIntExact(pageable.getOffset()));
         selectQuery.setMaxResults(pageable.getPageSize());
 
         List<Transaction> content = selectQuery.getResultList();
